requirements.txt
fyers-apiv2
flask
flask_ngrok
kiteconnect
selenium
nsetools
schedule
fastapi
bs4